requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
astroid==2.14.2
async-timeout==3.0.1
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.2
decorator==5.1.1
dill==0.3.6
discord.py==1.7.3
ffmpeg==1.4
frozenlist==1.3.3
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
platformdirs==3.1.0
pycparser==2.21
pylint==2.16.3
PyNaCl==1.5.0
pyspnego==0.8.0
python-dotenv==0.19.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
smbprotocol==1.10.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.5.0
urllib3==1.26.14
validators==0.20.0
wrapt==1.15.0
yarl==1.8.2
youtube-dl==2021.12.17
